+++ src/data/serialize_edge.cc
@@ -123,7 +123,7 @@ namespace neuroh5
                                     const vector<char> &recvbuf,
                                     const vector<int>& recvcounts,
                                     const vector<int>& rdispls,
-                                    const vector<uint32_t> &edge_attr_num,
+                                    const vector<size_t> &edge_attr_num,
                                     edge_map_t& prj_edge_map,
                                     size_t& num_unpacked_nodes,
                                     size_t& num_unpacked_edges
@@ -178,7 +178,7 @@ namespace neuroh5
 
     
     void deserialize_edge_map (const vector<char> &recvbuf,
-                               const vector<uint32_t> &edge_attr_num,
+                               const vector<size_t> &edge_attr_num,
                                edge_map_t& prj_edge_map,
                                size_t& num_unpacked_nodes,
                                size_t& num_unpacked_edges
